Add files via upload
New York City has published data on the SAT scores of students, along with additional demographic datasets.

New York City has a significant immigrant population, and is very diverse, so comparing demographic factors such as race, income, and gender with SAT scores is a good way to figure out if the SAT is a fair test. If certain racial groups consistently performed better on the SAT, we would have some evidence that the SAT is unfair, for example. 

We will clean and concat our various datasets before analyzing the data.