Add files via upload
A Project from scratch with two friends where we tried to create a device able to scan a product and tell us if it should be recycled or not based on the composition of the packaging. We found a dataset online with a lot of information about each product. The project is in French for the sake of our french teacher (sorry).